refactor: fusionne les dossiers domain et domaine
