+++ site/source/api/RechercheEntreprise/RechercheEntreprisesGouvFr.ts
@@ -1,10 +1,10 @@
-import { CodeActivite } from '@/domain/CodeActivite'
-import { CodeCatégorieJuridique } from '@/domain/CodeCatégorieJuridique'
-import { IsoDate, parseIsoDateString } from '@/domain/Date'
-import { Entreprise } from '@/domain/Entreprise'
-import { EntreprisesRepository } from '@/domain/EntreprisesRepository'
-import { Établissement } from '@/domain/Établissement'
-import { Siren, Siret } from '@/domain/Siren'
+import { CodeActivite } from '@/domaine/CodeActivite'
+import { CodeCatégorieJuridique } from '@/domaine/CodeCatégorieJuridique'
+import { IsoDate, parseIsoDateString } from '@/domaine/Date'
+import { Entreprise } from '@/domaine/Entreprise'
+import { EntreprisesRepository } from '@/domaine/EntreprisesRepository'
+import { Établissement } from '@/domaine/Établissement'
+import { Siren, Siret } from '@/domaine/Siren'
 
 /**
  * @see https://api.gouv.fr/documentation/api-recherche-entreprises
