+++ site/source/pages/simulateurs/comparaison-statuts/index.tsx
@@ -7,9 +7,9 @@ import { Emoji } from '@/design-system/emoji'
 import { Strong } from '@/design-system/typography'
 import { Link } from '@/design-system/typography/link'
 import { Body, Intro } from '@/design-system/typography/paragraphs'
-import { AssimiléSalariéContexte } from '@/domain/AssimiléSalariéContexte'
-import { AutoentrepreneurContexte } from '@/domain/AutoentrepreneurContexte'
-import { IndépendantContexte } from '@/domain/IndépendantContexte'
+import { AssimiléSalariéContexte } from '@/domaine/AssimiléSalariéContexte'
+import { AutoentrepreneurContexte } from '@/domaine/AutoentrepreneurContexte'
+import { IndépendantContexte } from '@/domaine/IndépendantContexte'
 import { useSitePaths } from '@/sitePaths'
 
 import Comparateur from './components/Comparateur'
